docker_flask_test/app/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
flask-restful
flask-migrate
cryptography
flask-mysqldb
gunicorn
flask_marshmallow
mysql-connector-python
pymysql
opencv-python
qrcode[pil]
opencv-python-headless
Pillow==10.0.0
